chat/requirements.txt
Django==3.2.5
channels==3.0.4
channels-redis==3.3.0
djangorestframework==3.12.4